chore(frontend): remove unused runtime interrupt export (#839)
